Adding a new upgrade
Adding a new upgrade to the code. This upgrade has been fully implented into the code, including the JavaScript part. Next I'll be working on adding code so then the upgrades will work. After that I'll be working on adding upgrades to the idea lister, and then fact holder, and file plot, which are the two hardest programs I will be working on in the entire FILEPLAN game I am working on. Check the wiki for more info.